Extract DirectBufferAccess to a new file
